pyproject.toml
[project]
name = "secret-store"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "William", email = "git@mxcz.fr" }
]
dependencies = [
    "paramiko>=3.4.0,<4",
    "pycryptodome==3.20.0",
    "pyhpke>=0.5.3",
]
readme = "README.md"
requires-python = ">= 3.10"

[tool.rye]
managed = true
dev-dependencies = []

[project.scripts]
secret-store = "secretstore.bin.cli:main"
